+++ Tickflo.Core/Services/Authentication/IPasswordHasher.cs
@@ -1,7 +1,8 @@
-namespace Tickflo.Core.Services.Auth;
+namespace Tickflo.Core.Services.Authentication;
 
 public interface IPasswordHasher
 {
     public string Hash(string input);
     public bool Verify(string input, string hash);
 }
+
