+++ Tickflo.Core/Services/Authentication/IPasswordSetupService.cs
@@ -1,6 +1,6 @@
 using Tickflo.Core.Entities;
 
-namespace Tickflo.Core.Services.Auth;
+namespace Tickflo.Core.Services.Authentication;
 
 public record TokenValidationResult(bool IsValid, string? ErrorMessage, int? UserId, string? UserEmail);
 public record SetPasswordResult(bool Success, string? ErrorMessage, string? LoginToken, string? WorkspaceSlug, int? UserId, string? UserEmail);
@@ -12,3 +12,5 @@ public interface IPasswordSetupService
     Task<SetPasswordResult> SetPasswordWithTokenAsync(string tokenValue, string newPassword);
     Task<SetPasswordResult> SetInitialPasswordAsync(int userId, string newPassword);
 }
+
+
