+++ Tickflo.Core/Services/Authentication/IWorkspaceRoleBootstrapper.cs
@@ -1,8 +1,10 @@
 using System.Threading.Tasks;
 
-namespace Tickflo.Core.Services.Auth;
+namespace Tickflo.Core.Services.Authentication;
 
 public interface IWorkspaceRoleBootstrapper
 {
     Task BootstrapAdminAsync(int workspaceId, int creatorUserId);
 }
+
+
