+++ Tickflo.Core/Services/Users/UserManagementService.cs
@@ -1,6 +1,6 @@
 using Tickflo.Core.Data;
 using Tickflo.Core.Entities;
-using Tickflo.Core.Services.Auth;
+using Tickflo.Core.Services.Authentication;
 
 namespace Tickflo.Core.Services.Users;
 
@@ -105,5 +105,6 @@ public class UserManagementService : IUserManagementService
         return null;
     }
 }
+
