+++ Tickflo.CoreTest/Services/Auth/PasswordSetupServiceTests.cs
@@ -1,7 +1,7 @@
 using Moq;
 using Tickflo.Core.Data;
 using Tickflo.Core.Entities;
-using Tickflo.Core.Services.Auth;
+using Tickflo.Core.Services.Authentication;
 using Xunit;
 
 namespace Tickflo.CoreTest.Services.Auth;
@@ -203,3 +203,4 @@ public class PasswordSetupServiceTests
             workspaceRepository);
     }
 }
+
