+++ Tickflo.Web/Pages/Signup.cshtml.cs
@@ -5,7 +5,7 @@ using Microsoft.AspNetCore.Mvc.RazorPages;
 using Tickflo.Core.Services;
 
 
-using Tickflo.Core.Services.Auth;
+using Tickflo.Core.Services.Authentication;
 namespace Tickflo.Web.Pages;
 
 [AllowAnonymous]
@@ -94,4 +94,5 @@ public class SignupInput
     [Compare("Password", ErrorMessage = "The password and confirmation password do not match.")]
     public string ConfirmPassword { get; set; } = "";
 }
+
