+++ Tickflo.Web/Pages/Users/Create.cshtml.cs
@@ -4,7 +4,7 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using Tickflo.Core.Data;
 using Tickflo.Core.Entities;
-using Tickflo.Core.Services.Auth;
+using Tickflo.Core.Services.Authentication;
 using Tickflo.Core.Services;
 
 using Tickflo.Core.Services.Common;
@@ -110,4 +110,5 @@ public class CreateModel : PageModel
         return RedirectToPage("/Workspace");
     }
 }
+
