+++ Tickflo.Web/Pages/Workspaces/UsersInvite.cshtml.cs
@@ -5,7 +5,7 @@ using Microsoft.AspNetCore.Http;
 using Tickflo.Core.Entities;
 using System.ComponentModel.DataAnnotations;
 using System.Security.Claims;
-using Tickflo.Core.Services.Auth;
+using Tickflo.Core.Services.Authentication;
 using Tickflo.Core.Services.Email;
 using Tickflo.Core.Utils;
 using System.Security.Cryptography;
@@ -139,4 +139,5 @@ public class UsersInviteModel : WorkspacePageModel
         return RedirectToPage("/Workspaces/Users", new { slug });
     }
 }
+
